webpack: refactor in to separate configs
- Moves certain webpack configs in to separate files to
  logically separate them
- Now we have a separate config for Base, Browser, and Vue
- webpack.config.js is the default name for the webpack
  cli to search, so it will be the main webpack file that
  collects all configs.

- Additionally, adds a type to stop errors for the CircularDependencyPlugin
